web: extract fixtures into a separate file
